# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Character API
    description: HTTP service for managing characters.
    version: ""
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /characters:
        get:
            tags:
                - character
            summary: getCharacters character
            operationId: character#getCharacters
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/CharacterResponse'
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/CharacterGetCharactersBadDataResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/CharacterGetCharactersIDNotFoundResponseBody'
                "507":
                    description: Insufficient Storage response.
                    schema:
                        $ref: '#/definitions/CharacterGetCharactersIDsExhaustedResponseBody'
            schemes:
                - http
        post:
            tags:
                - character
            summary: createCharacter character
            operationId: character#createCharacter
            parameters:
                - name: CreateCharacterRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CharacterCreateCharacterRequestBody'
                    required:
                        - name
                        - description
                        - health
                        - experience
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CharacterCreateCharacterResponseBody'
                        required:
                            - id
                            - name
                            - description
                            - health
                            - experience
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/CharacterCreateCharacterBadDataResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/CharacterCreateCharacterIDNotFoundResponseBody'
                "507":
                    description: Insufficient Storage response.
                    schema:
                        $ref: '#/definitions/CharacterCreateCharacterIDsExhaustedResponseBody'
            schemes:
                - http
    /characters/{id}:
        get:
            tags:
                - character
            summary: getCharacter character
            operationId: character#getCharacter
            parameters:
                - name: id
                  in: path
                  description: Character identifier
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CharacterGetCharacterResponseBody'
                        required:
                            - id
                            - name
                            - description
                            - health
                            - experience
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/CharacterGetCharacterBadDataResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/CharacterGetCharacterIDNotFoundResponseBody'
                "507":
                    description: Insufficient Storage response.
                    schema:
                        $ref: '#/definitions/CharacterGetCharacterIDsExhaustedResponseBody'
            schemes:
                - http
        put:
            tags:
                - character
            summary: updateCharacter character
            operationId: character#updateCharacter
            parameters:
                - name: id
                  in: path
                  description: Character identifier
                  required: true
                  type: integer
                - name: UpdateCharacterRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CharacterUpdateCharacterRequestBody'
                    required:
                        - name
                        - description
                        - health
                        - experience
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CharacterUpdateCharacterResponseBody'
                        required:
                            - id
                            - name
                            - description
                            - health
                            - experience
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/CharacterUpdateCharacterBadDataResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/CharacterUpdateCharacterIDNotFoundResponseBody'
                "507":
                    description: Insufficient Storage response.
                    schema:
                        $ref: '#/definitions/CharacterUpdateCharacterIDsExhaustedResponseBody'
            schemes:
                - http
        delete:
            tags:
                - character
            summary: deleteCharacter character
            operationId: character#deleteCharacter
            parameters:
                - name: id
                  in: path
                  description: Character identifier
                  required: true
                  type: integer
            responses:
                "204":
                    description: No Content response.
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/CharacterDeleteCharacterBadDataResponseBody'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/CharacterDeleteCharacterIDNotFoundResponseBody'
                "507":
                    description: Insufficient Storage response.
                    schema:
                        $ref: '#/definitions/CharacterDeleteCharacterIDsExhaustedResponseBody'
            schemes:
                - http
...
definitions:
    CharacterCreateCharacterBadDataResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: createCharacter_BadData_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    CharacterCreateCharacterIDNotFoundResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: createCharacter_IdNotFound_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    CharacterCreateCharacterIDsExhaustedResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: false
        description: createCharacter_IDsExhausted_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: false
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    CharacterCreateCharacterRequestBody:
        title: CharacterCreateCharacterRequestBody
        type: object
        properties:
            description:
                type: string
                description: Character description
                example: Architecto sapiente soluta.
            experience:
                type: integer
                description: Character experience
                example: 10916803485860537644
                format: int64
            health:
                type: integer
                description: Character health
                example: 14717078359636385948
                format: int64
            name:
                type: string
                description: Character name
                example: Placeat reprehenderit.
        example:
            description: Consequatur modi eos minima ea sunt et.
            experience: 1185649494524334896
            health: 10409843310894434344
            name: Laboriosam qui.
        required:
            - name
            - description
            - health
            - experience
    CharacterCreateCharacterResponseBody:
        title: CharacterCreateCharacterResponseBody
        type: object
        properties:
            description:
                type: string
                description: Character description
                example: Deleniti assumenda similique possimus atque architecto atque.
            experience:
                type: integer
                description: Character experience
                example: 18001780304994696442
                format: int64
            health:
                type: integer
                description: Character health
                example: 4970729370816610011
                format: int64
            id:
                type: integer
                description: Character identifier
                example: 8852108203923561610
                format: int64
            name:
                type: string
                description: Character name
                example: Nobis deleniti iste eligendi culpa molestias.
        example:
            description: Sapiente non.
            experience: 12998061760383796627
            health: 16966709443147470991
            id: 5906753977691565769
            name: Laboriosam dicta est modi.
        required:
            - id
            - name
            - description
            - health
            - experience
...
    CharacterGetCharactersBadDataResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: getCharacters_BadData_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    CharacterGetCharactersIDNotFoundResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: getCharacters_IdNotFound_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    CharacterGetCharactersIDsExhaustedResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: getCharacters_IDsExhausted_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    CharacterResponse:
        title: CharacterResponse
        type: object
        properties:
            description:
                type: string
                description: Character description
                example: Occaecati harum unde consectetur ut suscipit.
            experience:
                type: integer
                description: Character experience
                example: 4598434813898501559
                format: int64
            health:
                type: integer
                description: Character health
                example: 2225386145750905738
                format: int64
            id:
                type: integer
                description: Character identifier
                example: 8466815899249191702
                format: int64
            name:
                type: string
                description: Character name
                example: Ea modi est iste facilis unde.
        description: Character and attributes
        example:
            description: A ut alias qui.
            experience: 14328394898631220921
            health: 9521562689818229682
            id: 1684340369107676584
            name: Libero dolorem vero praesentium.
        required:
            - id
            - name
            - description
            - health
            - experience
...